requirements.txt
Flask
Flask-CORS
pandas
numpy
scikit-learn
joblib
openai
torch
pillow
torchvision